Moved all the code related to restoring wallet from file into separate class in order to be able to easily use it outside WalletActivity
