+++ wallet/src/de/schildbach/wallet/ui/RestoreFromFileHelper.java
@@ -1,0 +1,302 @@
+package de.schildbach.wallet.ui;
+
+import android.annotation.SuppressLint;
+import android.app.Activity;
+import android.app.Dialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.text.format.DateUtils;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.CheckBox;
+import android.widget.EditText;
+import android.widget.Spinner;
+import android.widget.TextView;
+
+import androidx.appcompat.app.AlertDialog;
+
+import com.google.common.base.Charsets;
+
+import org.bitcoinj.wallet.Wallet;
+import org.dash.wallet.common.Configuration;
+import org.dash.wallet.common.ui.DialogBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.BufferedReader;
+import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.LinkedList;
+import java.util.List;
+
+import de.schildbach.wallet.Constants;
+import de.schildbach.wallet.WalletApplication;
+import de.schildbach.wallet.util.Crypto;
+import de.schildbach.wallet.util.Io;
+import de.schildbach.wallet.util.WalletUtils;
+import de.schildbach.wallet_test.R;
+
+public class RestoreFromFileHelper {
+
+    protected static final Logger log = LoggerFactory.getLogger(RestoreFromFileHelper.class);
+
+
+    public static Dialog createRestoreWalletPermissionDialog(Context context) {
+        final DialogBuilder dialog = new DialogBuilder(context);
+        dialog.setTitle(R.string.restore_wallet_permission_dialog_title);
+        dialog.setMessage(context.getString(R.string.restore_wallet_permission_dialog_message));
+        dialog.singleDismissButton(null);
+        return dialog.create();
+    }
+
+    @SuppressLint("StringFormatInvalid")
+    public static Dialog createRestoreWalletDialog(final Activity activity, final OnRestoreWalletListener listener) {
+        final View view = activity.getLayoutInflater().inflate(R.layout.restore_wallet_dialog, null);
+        final TextView messageView = view.findViewById(R.id.restore_wallet_dialog_message);
+        final Spinner fileView = view.findViewById(R.id.import_keys_from_storage_file);
+        final EditText passwordView = view.findViewById(R.id.import_keys_from_storage_password);
+
+        final DialogBuilder dialog = new DialogBuilder(activity);
+        dialog.setTitle(R.string.import_keys_dialog_title);
+        dialog.setView(view);
+        dialog.setPositiveButton(R.string.import_keys_dialog_button_import, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(final DialogInterface dialog, final int which) {
+                final File file = (File) fileView.getSelectedItem();
+                final String password = passwordView.getText().toString().trim();
+                passwordView.setText(null); // get rid of it asap
+
+                if (WalletUtils.BACKUP_FILE_FILTER.accept(file))
+                    restoreWalletFromProtobuf(activity, file, listener);
+                else if (WalletUtils.KEYS_FILE_FILTER.accept(file))
+                    restorePrivateKeysFromBase58(activity, file, listener);
+                else if (Crypto.OPENSSL_FILE_FILTER.accept(file))
+                    restoreWalletFromEncrypted(activity, file, password, listener);
+            }
+        });
+        dialog.setNegativeButton(R.string.button_cancel, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(final DialogInterface dialog, final int which) {
+                passwordView.setText(null); // get rid of it asap
+            }
+        });
+        dialog.setOnCancelListener(new DialogInterface.OnCancelListener() {
+            @Override
+            public void onCancel(final DialogInterface dialog) {
+                passwordView.setText(null); // get rid of it asap
+            }
+        });
+
+        final FileAdapter adapter = new FileAdapter(activity) {
+            @SuppressLint("SetTextI18n")
+            @Override
+            public View getDropDownView(final int position, View row, final ViewGroup parent) {
+                final File file = getItem(position);
+                final boolean isExternal = Constants.Files.EXTERNAL_WALLET_BACKUP_DIR.equals(file.getParentFile());
+                final boolean isEncrypted = Crypto.OPENSSL_FILE_FILTER.accept(file);
+
+                if (row == null)
+                    row = inflater.inflate(R.layout.restore_wallet_file_row, parent, false);
+
+                final TextView filenameView = row.findViewById(R.id.wallet_import_keys_file_row_filename);
+                filenameView.setText(file.getName());
+
+                final TextView securityView = row.findViewById(R.id.wallet_import_keys_file_row_security);
+                final String encryptedStr = context
+                        .getString(isEncrypted ? R.string.import_keys_dialog_file_security_encrypted
+                                : R.string.import_keys_dialog_file_security_unencrypted);
+                final String storageStr = context
+                        .getString(isExternal ? R.string.import_keys_dialog_file_security_external
+                                : R.string.import_keys_dialog_file_security_internal);
+                securityView.setText(encryptedStr + ", " + storageStr);
+
+                final TextView createdView = row.findViewById(R.id.wallet_import_keys_file_row_created);
+                createdView.setText(context.getString(
+                        isExternal ? R.string.import_keys_dialog_file_created_manual
+                                : R.string.import_keys_dialog_file_created_automatic,
+                        DateUtils.getRelativeTimeSpanString(context, file.lastModified(), true)));
+
+                return row;
+            }
+        };
+
+        final String path;
+        final String backupPath = Constants.Files.EXTERNAL_WALLET_BACKUP_DIR.getAbsolutePath();
+        final String storagePath = Constants.Files.EXTERNAL_STORAGE_DIR.getAbsolutePath();
+        if (backupPath.startsWith(storagePath))
+            path = backupPath.substring(storagePath.length());
+        else
+            path = backupPath;
+        messageView.setText(activity.getString(R.string.import_keys_dialog_message, path));
+
+        fileView.setAdapter(adapter);
+
+        return dialog.create();
+    }
+
+    public static void prepareRestoreWalletDialog(Activity activity, boolean showReplaceWarning, final Dialog dialog) {
+        final AlertDialog alertDialog = (AlertDialog) dialog;
+
+        final List<File> files = new LinkedList<File>();
+
+        // external storage
+        final File[] externalFiles = Constants.Files.EXTERNAL_WALLET_BACKUP_DIR.listFiles();
+        if (externalFiles != null)
+            for (final File file : externalFiles)
+                if (Crypto.OPENSSL_FILE_FILTER.accept(file))
+                    files.add(file);
+
+        // internal storage
+        for (final String filename : activity.fileList())
+            if (filename.startsWith(Constants.Files.WALLET_KEY_BACKUP_PROTOBUF + '.'))
+                files.add(new File(activity.getFilesDir(), filename));
+
+        // sort
+        Collections.sort(files, new Comparator<File>() {
+            @Override
+            public int compare(final File lhs, final File rhs) {
+                return lhs.getName().compareToIgnoreCase(rhs.getName());
+            }
+        });
+
+        final View replaceWarningView = alertDialog.findViewById(R.id.restore_wallet_from_storage_dialog_replace_warning);
+        replaceWarningView.setVisibility(showReplaceWarning ? View.VISIBLE : View.GONE);
+
+        final Spinner fileView = (Spinner) alertDialog.findViewById(R.id.import_keys_from_storage_file);
+        final FileAdapter adapter = (FileAdapter) fileView.getAdapter();
+        adapter.setFiles(files);
+        fileView.setEnabled(!adapter.isEmpty());
+
+        final EditText passwordView = (EditText) alertDialog.findViewById(R.id.import_keys_from_storage_password);
+        passwordView.setText(null);
+
+        final ImportDialogButtonEnablerListener dialogButtonEnabler = new ImportDialogButtonEnablerListener(
+                passwordView, alertDialog) {
+            @Override
+            protected boolean hasFile() {
+                return fileView.getSelectedItem() != null;
+            }
+
+            @Override
+            protected boolean needsPassword() {
+                final File selectedFile = (File) fileView.getSelectedItem();
+                return selectedFile != null ? Crypto.OPENSSL_FILE_FILTER.accept(selectedFile) : false;
+            }
+        };
+        passwordView.addTextChangedListener(dialogButtonEnabler);
+        fileView.setOnItemSelectedListener(dialogButtonEnabler);
+
+        final CheckBox showView = (CheckBox) alertDialog.findViewById(R.id.import_keys_from_storage_show);
+        showView.setOnCheckedChangeListener(new ShowPasswordCheckListener(passwordView));
+    }
+
+    @SuppressLint("StringFormatInvalid")
+    private static void restoreWalletFromProtobuf(final Activity activity, final File file, final OnRestoreWalletListener listener) {
+        FileInputStream is = null;
+        try {
+            is = new FileInputStream(file);
+            listener.onRestoreWallet(WalletUtils.restoreWalletFromProtobuf(is, Constants.NETWORK_PARAMETERS));
+
+            log.info("successfully restored unencrypted wallet: {}", file);
+        } catch (final IOException x) {
+            final DialogBuilder dialog = DialogBuilder.warn(activity, R.string.import_export_keys_dialog_failure_title);
+            dialog.setMessage(activity.getString(R.string.import_keys_dialog_failure, x.getMessage()));
+            dialog.setPositiveButton(R.string.button_dismiss, null);
+            dialog.setNegativeButton(R.string.button_retry, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(final DialogInterface dialog, final int id) {
+                    listener.onRetryRequest();
+                }
+            });
+            dialog.show();
+
+            log.info("problem restoring unencrypted wallet: " + file, x);
+        } finally {
+            if (is != null) {
+                try {
+                    is.close();
+                } catch (final IOException x2) {
+                    // swallow
+                }
+            }
+        }
+    }
+
+    private static void restorePrivateKeysFromBase58(final Activity activity, final File file, final OnRestoreWalletListener listener) {
+        FileInputStream is = null;
+        try {
+            is = new FileInputStream(file);
+            listener.onRestoreWallet(WalletUtils.restorePrivateKeysFromBase58(is, Constants.NETWORK_PARAMETERS));
+            //remind user to backup since the key backup file does not have an HD seed
+            //Each time the user restores this backup file a new HD seed will be generated
+            Configuration config = ((WalletApplication) activity.getApplication()).getConfiguration();
+            config.armBackupReminder();
+            config.armBackupSeedReminder();
+            log.info("successfully restored unencrypted private keys: {}", file);
+        } catch (final IOException x) {
+            final DialogBuilder dialog = DialogBuilder.warn(activity, R.string.import_export_keys_dialog_failure_title);
+            dialog.setMessage(activity.getString(R.string.import_keys_dialog_failure, x.getMessage()));
+            dialog.setPositiveButton(R.string.button_dismiss, null);
+            dialog.setNegativeButton(R.string.button_retry, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(final DialogInterface dialog, final int id) {
+                    listener.onRetryRequest();
+                }
+            });
+            dialog.show();
+
+            log.info("problem restoring private keys: " + file, x);
+        } finally {
+            if (is != null) {
+                try {
+                    is.close();
+                } catch (final IOException x2) {
+                    // swallow
+                }
+            }
+        }
+    }
+
+    private static void restoreWalletFromEncrypted(final Activity activity, final File file, final String password, final OnRestoreWalletListener listener) {
+        try {
+            final BufferedReader cipherIn = new BufferedReader(
+                    new InputStreamReader(new FileInputStream(file), Charsets.UTF_8));
+            final StringBuilder cipherText = new StringBuilder();
+            Io.copy(cipherIn, cipherText, Constants.BACKUP_MAX_CHARS);
+            cipherIn.close();
+
+            final byte[] plainText = Crypto.decryptBytes(cipherText.toString(), password.toCharArray());
+            final InputStream is = new ByteArrayInputStream(plainText);
+
+            listener.onRestoreWallet(WalletUtils.restoreWalletFromProtobufOrBase58(is, Constants.NETWORK_PARAMETERS));
+
+            log.info("successfully restored encrypted wallet: {}", file);
+        } catch (final IOException x) {
+            final DialogBuilder dialog = DialogBuilder.warn(activity, R.string.import_export_keys_dialog_failure_title);
+            dialog.setMessage(activity.getString(R.string.import_keys_dialog_failure, x.getMessage()));
+            dialog.setPositiveButton(R.string.button_dismiss, null);
+            dialog.setNegativeButton(R.string.button_retry, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(final DialogInterface dialog, final int id) {
+                    listener.onRetryRequest();
+                }
+            });
+            dialog.show();
+
+            log.info("problem restoring wallet: " + file, x);
+        }
+    }
+
+    public interface OnRestoreWalletListener {
+
+        void onRestoreWallet(Wallet wallet);
+
+        void onRetryRequest();
+    }
+}
