+++ wallet/src/de/schildbach/wallet/ui/WalletActivity.java
@@ -25,7 +25,6 @@ import android.content.ClipDescription;
 import android.content.ClipboardManager;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.content.DialogInterface.OnCancelListener;
 import android.content.DialogInterface.OnClickListener;
 import android.content.Intent;
 import android.content.IntentFilter;
@@ -50,20 +49,13 @@ import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.widget.Toolbar;
 import android.telephony.TelephonyManager;
-import android.text.format.DateUtils;
 import android.view.ContextMenu;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.animation.AnimationUtils;
-import android.widget.CheckBox;
-import android.widget.EditText;
-import android.widget.Spinner;
-import android.widget.TextView;
 
-import com.google.common.base.Charsets;
 import com.google.common.collect.ImmutableList;
 import com.squareup.okhttp.HttpUrl;
 
@@ -73,25 +65,14 @@ import org.bitcoinj.core.VerificationException;
 import org.bitcoinj.core.VersionedChecksummedBytes;
 import org.bitcoinj.crypto.ChildNumber;
 import org.bitcoinj.wallet.Wallet;
-import org.bitcoinj.wallet.Wallet.BalanceType;
 import org.dash.wallet.common.Configuration;
 import org.dash.wallet.common.ui.DialogBuilder;
 import org.dash.wallet.integration.uphold.data.UpholdClient;
 import org.dash.wallet.integration.uphold.ui.UpholdAccountActivity;
 import org.dash.wallet.integration.uphold.ui.UpholdSplashActivity;
 
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.Currency;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Locale;
 
 import de.schildbach.wallet.Constants;
@@ -106,11 +87,8 @@ import de.schildbach.wallet.ui.send.SendCoinsActivity;
 import de.schildbach.wallet.ui.send.SweepWalletActivity;
 import de.schildbach.wallet.ui.widget.UpgradeWalletDisclaimerDialog;
 import de.schildbach.wallet.util.CrashReporter;
-import de.schildbach.wallet.util.Crypto;
 import de.schildbach.wallet.util.FingerprintHelper;
-import de.schildbach.wallet.util.Io;
 import de.schildbach.wallet.util.Nfc;
-import de.schildbach.wallet.util.WalletUtils;
 import de.schildbach.wallet_test.R;
 
 /**
@@ -647,152 +625,26 @@ public final class WalletActivity extends AbstractBindServiceActivity
     }
 
     private Dialog createRestoreWalletPermissionDialog() {
-        final DialogBuilder dialog = new DialogBuilder(this);
-        dialog.setTitle(R.string.restore_wallet_permission_dialog_title);
-        dialog.setMessage(getString(R.string.restore_wallet_permission_dialog_message));
-        dialog.singleDismissButton(null);
-        return dialog.create();
+        return RestoreFromFileHelper.createRestoreWalletPermissionDialog(this);
     }
 
     private Dialog createRestoreWalletDialog() {
-        final View view = getLayoutInflater().inflate(R.layout.restore_wallet_dialog, null);
-        final TextView messageView = (TextView) view.findViewById(R.id.restore_wallet_dialog_message);
-        final Spinner fileView = (Spinner) view.findViewById(R.id.import_keys_from_storage_file);
-        final EditText passwordView = (EditText) view.findViewById(R.id.import_keys_from_storage_password);
-
-        final DialogBuilder dialog = new DialogBuilder(this);
-        dialog.setTitle(R.string.import_keys_dialog_title);
-        dialog.setView(view);
-        dialog.setPositiveButton(R.string.import_keys_dialog_button_import, new OnClickListener() {
+        return RestoreFromFileHelper.createRestoreWalletDialog(this, new RestoreFromFileHelper.OnRestoreWalletListener() {
             @Override
-            public void onClick(final DialogInterface dialog, final int which) {
-                final File file = (File) fileView.getSelectedItem();
-                final String password = passwordView.getText().toString().trim();
-                passwordView.setText(null); // get rid of it asap
+            public void onRestoreWallet(Wallet wallet) {
+                restoreWallet(wallet);
+            }
 
-                if (WalletUtils.BACKUP_FILE_FILTER.accept(file))
-                    restoreWalletFromProtobuf(file);
-                else if (WalletUtils.KEYS_FILE_FILTER.accept(file))
-                    restorePrivateKeysFromBase58(file);
-                else if (Crypto.OPENSSL_FILE_FILTER.accept(file))
-                    restoreWalletFromEncrypted(file, password);
+            @Override
+            public void onRetryRequest() {
+                showDialog(DIALOG_RESTORE_WALLET);
             }
         });
-        dialog.setNegativeButton(R.string.button_cancel, new OnClickListener() {
-            @Override
-            public void onClick(final DialogInterface dialog, final int which) {
-                passwordView.setText(null); // get rid of it asap
-            }
-        });
-        dialog.setOnCancelListener(new OnCancelListener() {
-            @Override
-            public void onCancel(final DialogInterface dialog) {
-                passwordView.setText(null); // get rid of it asap
-            }
-        });
-
-        final FileAdapter adapter = new FileAdapter(this) {
-            @Override
-            public View getDropDownView(final int position, View row, final ViewGroup parent) {
-                final File file = getItem(position);
-                final boolean isExternal = Constants.Files.EXTERNAL_WALLET_BACKUP_DIR.equals(file.getParentFile());
-                final boolean isEncrypted = Crypto.OPENSSL_FILE_FILTER.accept(file);
-
-                if (row == null)
-                    row = inflater.inflate(R.layout.restore_wallet_file_row, null);
-
-                final TextView filenameView = (TextView) row.findViewById(R.id.wallet_import_keys_file_row_filename);
-                filenameView.setText(file.getName());
-
-                final TextView securityView = (TextView) row.findViewById(R.id.wallet_import_keys_file_row_security);
-                final String encryptedStr = context
-                        .getString(isEncrypted ? R.string.import_keys_dialog_file_security_encrypted
-                                : R.string.import_keys_dialog_file_security_unencrypted);
-                final String storageStr = context
-                        .getString(isExternal ? R.string.import_keys_dialog_file_security_external
-                                : R.string.import_keys_dialog_file_security_internal);
-                securityView.setText(encryptedStr + ", " + storageStr);
-
-                final TextView createdView = (TextView) row.findViewById(R.id.wallet_import_keys_file_row_created);
-                createdView.setText(context.getString(
-                        isExternal ? R.string.import_keys_dialog_file_created_manual
-                                : R.string.import_keys_dialog_file_created_automatic,
-                        DateUtils.getRelativeTimeSpanString(context, file.lastModified(), true)));
-
-                return row;
-            }
-        };
-
-        final String path;
-        final String backupPath = Constants.Files.EXTERNAL_WALLET_BACKUP_DIR.getAbsolutePath();
-        final String storagePath = Constants.Files.EXTERNAL_STORAGE_DIR.getAbsolutePath();
-        if (backupPath.startsWith(storagePath))
-            path = backupPath.substring(storagePath.length());
-        else
-            path = backupPath;
-        messageView.setText(getString(R.string.import_keys_dialog_message, path));
-
-        fileView.setAdapter(adapter);
-
-        return dialog.create();
     }
 
     private void prepareRestoreWalletDialog(final Dialog dialog) {
-        final AlertDialog alertDialog = (AlertDialog) dialog;
-
-        final List<File> files = new LinkedList<File>();
-
-        // external storage
-        final File[] externalFiles = Constants.Files.EXTERNAL_WALLET_BACKUP_DIR.listFiles();
-        if (externalFiles != null)
-            for (final File file : externalFiles)
-                if (Crypto.OPENSSL_FILE_FILTER.accept(file))
-                    files.add(file);
-
-        // internal storage
-        for (final String filename : fileList())
-            if (filename.startsWith(Constants.Files.WALLET_KEY_BACKUP_PROTOBUF + '.'))
-                files.add(new File(getFilesDir(), filename));
-
-        // sort
-        Collections.sort(files, new Comparator<File>() {
-            @Override
-            public int compare(final File lhs, final File rhs) {
-                return lhs.getName().compareToIgnoreCase(rhs.getName());
-            }
-        });
-
-        final View replaceWarningView = alertDialog
-                .findViewById(R.id.restore_wallet_from_storage_dialog_replace_warning);
-        final boolean hasCoins = wallet.getBalance(BalanceType.ESTIMATED).signum() > 0;
-        replaceWarningView.setVisibility(hasCoins ? View.VISIBLE : View.GONE);
-
-        final Spinner fileView = (Spinner) alertDialog.findViewById(R.id.import_keys_from_storage_file);
-        final FileAdapter adapter = (FileAdapter) fileView.getAdapter();
-        adapter.setFiles(files);
-        fileView.setEnabled(!adapter.isEmpty());
-
-        final EditText passwordView = (EditText) alertDialog.findViewById(R.id.import_keys_from_storage_password);
-        passwordView.setText(null);
-
-        final ImportDialogButtonEnablerListener dialogButtonEnabler = new ImportDialogButtonEnablerListener(
-                passwordView, alertDialog) {
-            @Override
-            protected boolean hasFile() {
-                return fileView.getSelectedItem() != null;
-            }
-
-            @Override
-            protected boolean needsPassword() {
-                final File selectedFile = (File) fileView.getSelectedItem();
-                return selectedFile != null ? Crypto.OPENSSL_FILE_FILTER.accept(selectedFile) : false;
-            }
-        };
-        passwordView.addTextChangedListener(dialogButtonEnabler);
-        fileView.setOnItemSelectedListener(dialogButtonEnabler);
-
-        final CheckBox showView = (CheckBox) alertDialog.findViewById(R.id.import_keys_from_storage_show);
-        showView.setOnCheckedChangeListener(new ShowPasswordCheckListener(passwordView));
+        final boolean hasCoins = wallet.getBalance(Wallet.BalanceType.ESTIMATED).signum() > 0;
+        RestoreFromFileHelper.prepareRestoreWalletDialog(this, hasCoins, dialog);
     }
 
     private void showRestoreWalletFromSeedDialog() {
@@ -994,102 +846,7 @@ public final class WalletActivity extends AbstractBindServiceActivity
         return dialog.create();
     }
 
-    private void restoreWalletFromEncrypted(final File file, final String password) {
-        try {
-            final BufferedReader cipherIn = new BufferedReader(
-                    new InputStreamReader(new FileInputStream(file), Charsets.UTF_8));
-            final StringBuilder cipherText = new StringBuilder();
-            Io.copy(cipherIn, cipherText, Constants.BACKUP_MAX_CHARS);
-            cipherIn.close();
-
-            final byte[] plainText = Crypto.decryptBytes(cipherText.toString(), password.toCharArray());
-            final InputStream is = new ByteArrayInputStream(plainText);
-
-            restoreWallet(WalletUtils.restoreWalletFromProtobufOrBase58(is, Constants.NETWORK_PARAMETERS));
-
-            log.info("successfully restored encrypted wallet: {}", file);
-        } catch (final IOException x) {
-            final DialogBuilder dialog = DialogBuilder.warn(this, R.string.import_export_keys_dialog_failure_title);
-            dialog.setMessage(getString(R.string.import_keys_dialog_failure, x.getMessage()));
-            dialog.setPositiveButton(R.string.button_dismiss, null);
-            dialog.setNegativeButton(R.string.button_retry, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(final DialogInterface dialog, final int id) {
-                    showDialog(DIALOG_RESTORE_WALLET);
-                }
-            });
-            dialog.show();
-
-            log.info("problem restoring wallet: " + file, x);
-        }
-    }
-
-    private void restoreWalletFromProtobuf(final File file) {
-        FileInputStream is = null;
-        try {
-            is = new FileInputStream(file);
-            restoreWallet(WalletUtils.restoreWalletFromProtobuf(is, Constants.NETWORK_PARAMETERS));
-
-            log.info("successfully restored unencrypted wallet: {}", file);
-        } catch (final IOException x) {
-            final DialogBuilder dialog = DialogBuilder.warn(this, R.string.import_export_keys_dialog_failure_title);
-            dialog.setMessage(getString(R.string.import_keys_dialog_failure, x.getMessage()));
-            dialog.setPositiveButton(R.string.button_dismiss, null);
-            dialog.setNegativeButton(R.string.button_retry, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(final DialogInterface dialog, final int id) {
-                    showDialog(DIALOG_RESTORE_WALLET);
-                }
-            });
-            dialog.show();
-
-            log.info("problem restoring unencrypted wallet: " + file, x);
-        } finally {
-            if (is != null) {
-                try {
-                    is.close();
-                } catch (final IOException x2) {
-                    // swallow
-                }
-            }
-        }
-    }
-
-    private void restorePrivateKeysFromBase58(final File file) {
-        FileInputStream is = null;
-        try {
-            is = new FileInputStream(file);
-            restoreWallet(WalletUtils.restorePrivateKeysFromBase58(is, Constants.NETWORK_PARAMETERS));
-            //remind user to backup since the key backup file does not have an HD seed
-            //Each time the user restores this backup file a new HD seed will be generated
-            config.armBackupReminder();
-            config.armBackupSeedReminder();
-            log.info("successfully restored unencrypted private keys: {}", file);
-        } catch (final IOException x) {
-            final DialogBuilder dialog = DialogBuilder.warn(this, R.string.import_export_keys_dialog_failure_title);
-            dialog.setMessage(getString(R.string.import_keys_dialog_failure, x.getMessage()));
-            dialog.setPositiveButton(R.string.button_dismiss, null);
-            dialog.setNegativeButton(R.string.button_retry, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(final DialogInterface dialog, final int id) {
-                    showDialog(DIALOG_RESTORE_WALLET);
-                }
-            });
-            dialog.show();
-
-            log.info("problem restoring private keys: " + file, x);
-        } finally {
-            if (is != null) {
-                try {
-                    is.close();
-                } catch (final IOException x2) {
-                    // swallow
-                }
-            }
-        }
-    }
-
-    public void restoreWallet(final Wallet wallet) throws IOException {
+    public void restoreWallet(final Wallet wallet) {
         application.replaceWallet(wallet);
         getSharedPreferences(Constants.WALLET_LOCK_PREFS_NAME, Context.MODE_PRIVATE).edit().clear().commit();
